refactor: fix deps, make use-get-layer-config for page-specific layerconfig
